Fixed hooks in x64
Everything beside DoCallScript works, and unit tests pass (excluding the Script Array test)

+++ YYToolkit/Src/Core/Features/API/Internal.hpp
@@ -6,7 +6,7 @@
 struct CModule
 {
 	uintptr_t Base;
-	uintptr_t Size;
+	DWORD Size;
 	uintptr_t EntryPoint;
 };
